refactor: pattern label + summary table + index result badge helpers
