+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_details/indices_details/pattern/pattern_summary/pattern_label/index.tsx
@@ -8,11 +8,12 @@
 import { EuiTitle, EuiToolTip, EuiFlexGroup, EuiFlexItem } from '@elastic/eui';
 import React from 'react';
 
-import { getPatternResultTooltip, showResult } from './helpers';
-import { IlmPhaseCounts } from './ilm_phase_counts';
-import * as i18n from '../translations';
 import type { IlmExplainPhaseCounts } from '../../../../../types';
 import { IndexResultBadge } from '../../index_result_badge';
+import * as i18n from '../translations';
+import { IlmPhaseCounts } from './ilm_phase_counts';
+import { getPatternResultTooltip } from './utils/get_pattern_result_tooltip';
+import { showResult } from './utils/show_result';
 
 interface Props {
   incompatible: number | undefined;
