+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_details/indices_details/pattern/pattern_summary/pattern_label/utils/show_result.ts
@@ -1,0 +1,20 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+interface ShowResultProps<T> {
+  incompatible: T;
+  indices: T;
+  indicesChecked: T;
+}
+
+export const showResult = <T extends number | undefined>(
+  opts: ShowResultProps<T>
+): opts is ShowResultProps<NonNullable<T>> =>
+  opts.incompatible != null &&
+  opts.indices != null &&
+  opts.indicesChecked != null &&
+  opts.indices === opts.indicesChecked;
