+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_details/indices_details/pattern/summary_table/translations.ts
@@ -35,13 +35,6 @@ export const EXPAND_ROWS = i18n.translate(
   }
 );
 
-export const FAILED = i18n.translate(
-  'securitySolutionPackages.ecsDataQualityDashboard.summaryTable.failedTooltip',
-  {
-    defaultMessage: 'Failed',
-  }
-);
-
 export const ILM_PHASE = i18n.translate(
   'securitySolutionPackages.ecsDataQualityDashboard.summaryTable.ilmPhaseColumn',
   {
@@ -90,13 +83,6 @@ export const INDEX_TOOL_TIP = (pattern: string) =>
     defaultMessage: 'This index matches the pattern or index name: {pattern}',
   });
 
-export const PASSED = i18n.translate(
-  'securitySolutionPackages.ecsDataQualityDashboard.summaryTable.passedTooltip',
-  {
-    defaultMessage: 'Passed',
-  }
-);
-
 export const RESULT = i18n.translate(
   'securitySolutionPackages.ecsDataQualityDashboard.summaryTable.resultColumn',
   {
@@ -115,13 +101,6 @@ export const LAST_CHECK = i18n.translate(
   'securitySolutionPackages.ecsDataQualityDashboard.summaryTable.lastCheckColumn',
   {
     defaultMessage: 'Last check',
-  }
-);
-
-export const THIS_INDEX_HAS_NOT_BEEN_CHECKED = i18n.translate(
-  'securitySolutionPackages.ecsDataQualityDashboard.summaryTable.thisIndexHasNotBeenCheckedTooltip',
-  {
-    defaultMessage: 'This index has not been checked',
   }
 );
