+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_details/indices_details/pattern/translations.ts
@@ -22,3 +22,24 @@ export const LOADING_STATS = i18n.translate(
     defaultMessage: 'Loading stats',
   }
 );
+
+export const PASSED = i18n.translate(
+  'securitySolutionPackages.ecsDataQualityDashboard.passedTooltip',
+  {
+    defaultMessage: 'Passed',
+  }
+);
+
+export const FAILED = i18n.translate(
+  'securitySolutionPackages.ecsDataQualityDashboard.failedTooltip',
+  {
+    defaultMessage: 'Failed',
+  }
+);
+
+export const THIS_INDEX_HAS_NOT_BEEN_CHECKED = i18n.translate(
+  'securitySolutionPackages.ecsDataQualityDashboard.thisIndexHasNotBeenCheckedTooltip',
+  {
+    defaultMessage: 'This index has not been checked',
+  }
+);
